opengl的编译问题还没解决
一直提示找不到glGenBuffers，但实际上glew中是存在的

+++ 3dfeature/Curvature.hh
@@ -219,7 +219,7 @@ void discrete_mean_curv_op( const MeshT&                        _m,
     }
   }
 
-  _n /= 2.0*_area;
+  _n /= (float)(2.0*_area);
 }
 
 template <typename T>
